+++ propagators/baggage.go
@@ -12,14 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package baggage
+package propagators
 
 import (
 	"context"
 	"net/url"
 	"strings"
 
-	"go.opentelemetry.io/otel/api/propagation"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/internal/baggage"
 	"go.opentelemetry.io/otel/label"
 )
 
@@ -27,22 +28,17 @@ import (
 // https://github.com/open-telemetry/opentelemetry-specification/blob/18b2752ebe6c7f0cdd8c7b2bcbdceb0ae3f5ad95/specification/correlationcontext/api.md#header-name
 const baggageHeader = "otcorrelations"
 
-// Baggage propagates Key:Values in W3C CorrelationContext
-// format.
-// nolint:golint
+// Baggage is a propagator that supports the W3C Baggage format.
+//
+// This propagates user-defined baggage associated with a trace. The complete
+// specification is defined at https://w3c.github.io/baggage/.
 type Baggage struct{}
 
-var _ propagation.HTTPPropagator = Baggage{}
+var _ otel.TextMapPropagator = Baggage{}
 
-// DefaultHTTPPropagator returns the default context correlation HTTP
-// propagator.
-func DefaultHTTPPropagator() propagation.HTTPPropagator {
-	return Baggage{}
-}
-
-// Inject implements HTTPInjector.
-func (b Baggage) Inject(ctx context.Context, supplier propagation.HTTPSupplier) {
-	baggageMap := MapFromContext(ctx)
+// Inject sets baggage key-values from ctx into the carrier.
+func (b Baggage) Inject(ctx context.Context, carrier otel.TextMapCarrier) {
+	baggageMap := baggage.MapFromContext(ctx)
 	firstIter := true
 	var headerValueBuilder strings.Builder
 	baggageMap.Foreach(func(kv label.KeyValue) bool {
@@ -57,18 +53,18 @@ func (b Baggage) Inject(ctx context.Context, supplier propagation.HTTPSupplier) 
 	})
 	if headerValueBuilder.Len() > 0 {
 		headerString := headerValueBuilder.String()
-		supplier.Set(baggageHeader, headerString)
+		carrier.Set(baggageHeader, headerString)
 	}
 }
 
-// Extract implements HTTPExtractor.
-func (b Baggage) Extract(ctx context.Context, supplier propagation.HTTPSupplier) context.Context {
-	baggage := supplier.Get(baggageHeader)
-	if baggage == "" {
-		return ctx
+// Extract returns a copy of parent with the baggage from the carrier added.
+func (b Baggage) Extract(parent context.Context, carrier otel.TextMapCarrier) context.Context {
+	bVal := carrier.Get(baggageHeader)
+	if bVal == "" {
+		return parent
 	}
 
-	baggageValues := strings.Split(baggage, ",")
+	baggageValues := strings.Split(bVal, ",")
 	keyValues := make([]label.KeyValue, 0, len(baggageValues))
 	for _, baggageValue := range baggageValues {
 		valueAndProps := strings.Split(baggageValue, ";")
@@ -104,15 +100,15 @@ func (b Baggage) Extract(ctx context.Context, supplier propagation.HTTPSupplier)
 
 	if len(keyValues) > 0 {
 		// Only update the context if valid values were found
-		return ContextWithMap(ctx, NewMap(MapUpdate{
+		return baggage.ContextWithMap(parent, baggage.NewMap(baggage.MapUpdate{
 			MultiKV: keyValues,
 		}))
 	}
 
-	return ctx
+	return parent
 }
 
-// GetAllKeys implements HTTPPropagator.
-func (b Baggage) GetAllKeys() []string {
+// Fields returns the keys who's values are set with Inject.
+func (b Baggage) Fields() []string {
 	return []string{baggageHeader}
 }